Added indicators to ensure that the tests have been properly torn down in the past

+++ jest-global-teardown.js
@@ -1,16 +1,36 @@
 import fs from "fs";
 import path from "path";
+import { fileURLToPath } from "url";
 
 export default async function globalTeardown() {
 	// TODO: Tear down the test environment in its entirety
-	const nonceDir = global.nonceDir;
-	const fileList = fs.readdirSync(nonceDir);
-	if(fileList.length !== 1) {
-		console.warn(`Global teardown expected only 1 file from global setup in ${nonceDir} but found ${fileList.length}`);
-		console.warn("This implies that Jest has been run before and interrupted and did not fully teardown in the past");
-		console.warn(`Ensure that your environment has been successfully wiped or rerun teardown alone a couple times and remove the extra files from ${nonceDir}`);
+	if(global.setupUsed) {
+		console.log(`Removing Nonce File from ${global.nonceFile}`);
+		fs.rmSync(global.nonceFile);
 	}
+}
 
-	console.log(`Removing Nonce File from ${global.nonceFile}`);
-	fs.rmSync(global.nonceFile);
+// Run teardown directly if running directly from command line with `node jest-global-teardown.js`
+if(path.resolve(process.argv[1]) === path.resolve(fileURLToPath(import.meta.url))) {
+	const __filename = fileURLToPath(import.meta.url);
+	const __dirname = path.dirname(__filename);
+	const nonceDir = path.resolve(__dirname, "private/nonce");
+
+	console.log("Beginning Teardown ONLY");
+
+	globalTeardown()
+		.then(() => {
+			console.log("Teardown Complete");
+			console.log("Wiping contents of nonce directory so that setup doesn't complain about teardown not being run next time");
+
+			const nonceContents = fs.readdirSync(nonceDir);
+			for(const nonceFile of nonceContents) {
+				fs.rmSync(path.resolve(nonceDir, nonceFile));
+			}
+			console.log("Successfully wiped nonce directory");
+		})
+		.catch(err => {
+			console.warn("Teardown failed:");
+			console.error(err);
+		});
 }
